app/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
requests==2.32.3
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.2.2
rq==1.16.2
redis==5.0.8
jinja2==3.1.4
python-multipart==0.0.9
